% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ rematch }
\alias{ rematch }
\alias{ rematch }
\title{ Match Regular Expressions with a Nicer 'API' }
\description{
A small wrapper on 'regexpr' to extract the matches and captured groups from
  the match of a regular expression to a character vector.
}
