% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{re_match_all}
\alias{re_match_all}
\title{Extract all matches of a regular expression}
\usage{
re_match_all(text, pattern, ...)
}
\arguments{
\item{text}{Character vector.}

\item{pattern}{Regular expression, defaults to be a PCRE
expression. See \code{\link[base]{regex}} for more about
regular expressions.}

\item{...}{Additional arguments to pass to
\code{\link[base]{regexpr}}.}
}
\description{
This function is a thin wrapper on the \code{\link[base]{gregexpr}}
base R function, to provide an API that is easier to use. It is
similar to \code{\link{re_match}}, but extracts all matches, including
potentially named capture groups.
}

