% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/rematch2-package.R
\docType{package}
\name{rematch2-package}
\alias{rematch2}
\alias{rematch2-package}
\title{Match Regular Expressions with a Nicer 'API'}
\description{
A small wrapper on 'regexpr' to extract the matches and captured
groups from the match of a regular expression to a character vector.
See \code{\link[=re_match]{re_match()}}.

Wrappers on 'regexpr' and 'gregexpr' to return the match results in tidy data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/rematch2#readme}
  \item Report bugs at \url{https://github.com/r-lib/rematch2/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/r-lib/rematch2#readme}
  \item Report bugs at \url{https://github.com/r-lib/rematch2/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Matthew Lincoln \email{matthew.d.lincoln@gmail.com} [contributor]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
